requirements.txt
absl-py 
aiohttp==3.7.4.post0
aiosignal 
astunparse==1.6.3
async-timeout==3.0.1
asynctest==0.13.0
attrs 
bardapi==0.1.4
blinker==1.4
bnlp-toolkit==3.3.0
bntransformer==2.1.0
Bottleneck 
brotlipy==0.7.0
cachetools 
certifi 
cffi 
chardet==4.0.0
charset-normalizer 
click==8.1.3
cryptography 
emoji==1.7.0
fasttext==0.9.2
filelock==3.10.0
flatbuffers 
flit_core 
frozenlist 
ftfy==6.1.1
gast 
gensim==4.0.1
google-auth 
google-auth-oauthlib 
google-pasta 
GoogleBard==0.0.7
grpcio 
h5py 
huggingface-hub==0.0.8
idna 
importlib-metadata==6.0.0
joblib==1.2.0
keras 
Keras-Preprocessing 
laserembeddings==1.1.2
Levenshtein==0.20.9
Markdown 
markdown-it-py==2.2.0
MarkupSafe 
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random 
mkl-service==2.4.0
mock==5.0.1
multidict 
nltk==3.8.1
numexpr 
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib 
openai==0.27.4
opt-einsum 
packaging 
pandas==1.3.5
Pillow==9.4.0
prompt-toolkit==3.0.38
protobuf==3.20.3
py-cord==1.7.3
pyasn1 
pyasn1-modules==0.2.8
pybind11==2.10.3
pycparser 
Pygments==2.15.0
PyJWT 
pyOpenSSL 
PySocks 
python-crfsuite==0.9.9
python-dateutil 
python-dotenv==0.21.1
python-Levenshtein==0.20.9
pytz 
PyYAML==6.0
rapidfuzz==2.13.7
regex==2022.10.31
requests 
requests-oauthlib==1.3.0
rich==13.3.4
rsa 
sacremoses==0.0.35
scikit-learn==1.0.2
scipy 
sentence-transformers==2.2.2
sentencepiece==0.1.97
six 
sklearn-crfsuite==0.3.6
smart-open==6.3.0
subword-nmt==0.3.8
tabulate==0.9.0
tensorboard 
tensorboard-data-server 
tensorboard-plugin-wit 
tensorflow 
tensorflow-estimator 
tensorflow-hub==0.12.0
termcolor 
threadpoolctl==3.1.0
tokenizers==0.10.3
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.6.1
transliterate==1.10.2
typing_extensions==4.5.0
urllib3 
wcwidth==0.2.6
Werkzeug 
wrapt 
yarl 
zipp==3.15.0
